Move transaction hooks that don't fetch data out of daoData directory
