+++ src/components/DAOs/DAO/Proposals/ProposalQueue.tsx
@@ -1,5 +1,5 @@
 import { ProposalData } from "../../../../contexts/daoData/useProposals";
-import useQueueTransaction from "../../../../contexts/daoData/useQueueTransaction";
+import useQueueTransaction from "../../../../hooks/useQueueTransaction";
 import { PrimaryButton } from "../../../ui/forms/Button";
 
 function ProposalQueue({ proposal }: { proposal: ProposalData }) {
