+++ src/components/DAOs/New/index.tsx
@@ -2,11 +2,11 @@ import { useState } from "react";
 import DAODetails from "./DAODetails";
 import TokenDetails from "./TokenDetails";
 import GovernanceDetails from "./GovernanceDetails";
-import useDeployDAO from "../../../contexts/daoData/useDeployDAO";
+import useDeployDAO from "../../../hooks/useDeployDAO";
 import ContentBox from "../../ui/ContentBox";
 import LeftArrow from "../../ui/svg/LeftArrow";
 import RightArrow from "../../ui/svg/RightArrow";
-import { TokenAllocation } from "../../../contexts/daoData/useDeployDAO";
+import { TokenAllocation } from "../../../hooks/useDeployDAO";
 import { SecondaryButton, TextButton, PrimaryButton } from "../../ui/forms/Button";
 import H1 from "../../ui/H1";
 import ConnectWalletToast from "../shared/ConnectWalletToast";
