+++ src/hooks/useCreateProposal.ts
@@ -1,10 +1,10 @@
 import { useCallback, useEffect } from 'react'
-import { useTransaction } from '../web3Data/transactions';
-import { useWeb3 } from '../web3Data';
+import { useTransaction } from '../contexts/web3Data/transactions';
+import { useWeb3 } from '../contexts/web3Data';
 import { BigNumber } from 'ethers';
 import { useNavigate } from 'react-router';
-import { GovernorModule, GovernorModule__factory } from '../../typechain-types';
-import { useDAOData } from './index';
+import { GovernorModule, GovernorModule__factory } from '../typechain-types';
+import { useDAOData } from '../contexts/daoData/index';
 
 type ProposalData = {
   targets: string[];
